refactor: extract TestRailwayNetwork to helpers module
Address code review comments on PR #108:

1. Fix test module import issue:
   - Move TestRailwayNetwork class from conftest.py to
     tests/helpers/railway_network.py to avoid pytest discovery issues
   - Update import in test_test_network.py to use helpers module
   - This fixes linter warning about importing test modules

2. Improve codebase organization:
   - Reduce conftest.py from 1631 lines to 535 lines (~1100 lines moved)
   - TestRailwayNetwork now in dedicated module for better maintainability
   - Backward compatibility maintained with import in conftest.py

Changes:
- backend/tests/conftest.py: Remove TestRailwayNetwork class, add import
- backend/tests/helpers/railway_network.py: New module with TestRailwayNetwork
- backend/tests/test_test_network.py: Update import statement

All 802 tests pass, 98.44% coverage maintained.

Addresses review comment: "Don't import test modules" (dont-import-test-modules)
Addresses review comment: "Consider breaking up the massive TestRailwayNetwork class"
