+++ backend/tests/test_test_network.py
@@ -1,6 +1,6 @@
 """Unit tests for TestRailwayNetwork factory methods."""
 
-from tests.conftest import TestRailwayNetwork
+from tests.helpers.railway_network import TestRailwayNetwork
 
 
 class TestHubChildrenFactories:
